* refactor to support null safety ** switch to audioplayers package from audioplayer as part of process to accomplish this ** switch to Get package instead of using stacked_services package as part of process to accomplish this

+++ client/lib/infrastructure/favorited_words/favorited_words_local_data_source.dart
@@ -1,5 +1,4 @@
 import 'package:dartz/dartz.dart';
-import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
 import 'package:kt_dart/collection.dart';
 
@@ -12,14 +11,19 @@ import '../core/dtos/word_search_dto.dart';
 class FavoritedWordsLocalDataSource implements IFavoritedWordsLocalDataSource {
   final WordSearchDao wordSearchDao;
 
-  FavoritedWordsLocalDataSource({@required this.wordSearchDao});
+  FavoritedWordsLocalDataSource({required this.wordSearchDao});
 
   @override
   Future<Unit> addFavoritedSearch(WordSearchDto search) async {
     try {
       final existingOrNullWordSearch = await wordSearchDao.findById(search.word.id);
       if (existingOrNullWordSearch == null) {
-        await wordSearchDao.insert(WordSearchDto(word: search.word, isFavorited: true));
+        await wordSearchDao.insert(WordSearchDto(
+          word: search.word,
+          isFavorited: true,
+          results: search.results,
+        ));
+
         return unit;
       } else {
         final updatedObj = existingOrNullWordSearch.copyWith(isFavorited: true);
@@ -35,11 +39,16 @@ class FavoritedWordsLocalDataSource implements IFavoritedWordsLocalDataSource {
   Future<Unit> deleteFavoritedSearch(WordSearchDto search) async {
     try {
       final existingSearch = await wordSearchDao.findById(search.word.id);
+
+      if (existingSearch == null) {
+        throw const FavoritedWordsException.localDatabaseProcessingException();
+      }
+
       if (existingSearch.lastSearchedAt == null) {
         await wordSearchDao.delete(existingSearch);
         return unit;
       } else {
-        final updatedSearch = existingSearch.copyWith(isFavorited: null);
+        final updatedSearch = existingSearch.copyWith(isFavorited: false);
         await wordSearchDao.update(updatedSearch);
         return unit;
       }
